training/requirements.txt
# Multi-Agent MCTS Training Pipeline Requirements

# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
peft>=0.4.0

# Sentence Embeddings and RAG
sentence-transformers>=2.2.0
pinecone>=5.0.0  # Serverless Pinecone vector database
rank-bm25>=0.2.2
# faiss-cpu>=1.7.4  # Legacy: use faiss-gpu for GPU support

# Experiment Tracking
wandb>=0.15.0
mlflow>=2.5.0

# Configuration and Data
pyyaml>=6.0
omegaconf>=2.3.0
pydantic>=2.0.0

# Numerical and Scientific
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Monitoring and Profiling
psutil>=5.9.0
tensorboard>=2.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
hypothesis>=6.82.0

# Utilities
tqdm>=4.65.0
colorlog>=6.7.0
python-dotenv>=1.0.0

# Code corpus builder
PyGithub>=2.1.1  # GitHub API access

# Knowledge Graph
networkx>=3.0  # Graph data structure and algorithms
openai>=1.0.0  # LLM for knowledge extraction
neo4j>=5.0.0  # Optional: Neo4j graph database driver

# Research corpus builder
arxiv>=2.0.0  # arXiv API client
requests>=2.31.0  # HTTP requests

# Optional: Advanced Features
# chromadb>=0.4.0  # Alternative to FAISS
# langchain>=0.0.300  # For LangGraph integration
# opentelemetry-api>=1.20.0  # Distributed tracing
